src/main/java/Car.java
/* A class that implements the Tradable and Drivable interfaces.
 */
public class Car implements Drivable, Tradable{
    private int maxSpeed;

    public Car() {
        this.maxSpeed = 3;
    }

    @Override
    public void upgradeSpeed() {
        this.maxSpeed++;
    }

    @Override
    public void downgradeSpeed() {
        this.maxSpeed--;
    }

    @Override
    public int getMaxSpeed() {
        return this.maxSpeed;
    }

    @Override
    public int getPrice() {
        return 30;
    }

    @Override
    public String toString(){
        String id = super.toString();
        return id + " (Max speed: " + this.maxSpeed +")";
    }
}
